\name{fume-package}
\alias{fume-package}
\alias{fume}
\docType{package}
\title{
A set of functions for fire danger and climate analysis
}
\description{
This package includes functions for climate fire danger estimation, trend analysis and RCM bias correction and an example dataset of observations and reanalysis data from Spain. 

This package was developed for the Course on Forest Fires in the Perspective of Global Change, held in Zaragoza (Spain) the 13th to 17th February 2012 in the context of FUME Project, which has received funding from the European Union's Seventh Framework Programme (FP7/2007--2013) under grant agreement 243888 (FUME Project).
}
\details{
\tabular{ll}{
Package: \tab fume\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-02-11\cr
License: \tab GPL-2 \cr
LazyLoad: \tab yes\cr
}
The package provides functions for Canadian Fire Weather index calculations, the modified Mann-Kendall trend test for serially correlated data and a function containing different implementation for RCM/GCM bias correction.
}
\author{
Santander Meteorology Group

Maintainer: Joaquin Bedia <joaquin.bedia@unican.es>
}
\references{
Amengual, A., Homar, V., Romero, R., Alonso, S. and Ramis, C. A statistical adjustment of regional climate model outputs to local scales: Application to Platja de Palma, Spain. Journal of Climate. In press.

Bedia, J., Herrera, S., Gutierrez, J.M., Zavala, G., Urbieta, I.R. and Moreno, J.M. Sensitivity of Fire Weather Index to different reamalysis products in the Iberian Peninsula. Natural Hazards and Earth System Sciences. In press.

Hamed, K.H. and A.R. Rao, (1998). A modified Mann Kendall trend test for autocorrelated data. Journal of Hydrology 204, 182-196.

van Wagner, C.E. and T.L. Pickett (1985). Equations and FORTRAN program for the Canadian Forest Fire Weather Index system. Canadian Forestry Service Tech. Rep. 33. Ottawa, Canada.
}
\keyword{ package }
