\name{lappend}
\alias{lappend}

\title{
Adds a new element to a list
}
\description{
The function basically adds a new element to a list. It is used internally
}
\usage{
lappend(lst, obj)
}

\arguments{
  \item{lst}{
A list
}
  \item{obj}{
An element to be added to the existing list
}
}
\details{
This function is used internally by other functions within this package
}
\value{
A list containing the new element in its last position 

}

\author{
Santander Meteorology Group
<\url{http://www.meteo.unican.es}>
}
